requirements.txt
scp
paramiko
libvirt-python
python-telegram-bot
pyotp
